+++ odoo/local-src/sf_mrp/models/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of sensefly.
-import mrp_workorder
+from . import mrp_workorder
+from . import mrp_production